Perubahan UI

+++ tasks/views.py
@@ -1,30 +1,30 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import Task
 
 def index(request):
     query = request.GET.get('search')
-    if query :
+    if query:
         tasks = Task.objects.filter(title__icontains=query)
-    else :
+    else:
         tasks = Task.objects.all()
-    
-    # Cek jika request adalah POST untuk menambah tugas baru
+
     if request.method == 'POST':
         title = request.POST.get('title')
-        if title:  # Pastikan title tidak kosong
+        if title:
             Task.objects.create(title=title)
-            return redirect('index')  # Redirect ke halaman index
+            return redirect('index')
 
     return render(request, 'tasks/index.html', {'tasks': tasks, 'search_query': query})
 
-# Membuat fungsi complate tsak untuk melakukan perubahan pada UI index
-def complete_task(request, task_id):
-    task = Task.objects.get(id=task_id)
-    task.completed = True
-    task.save()
-    return redirect('index')
+def edit_task(request, task_id):
+    task = get_object_or_404(Task, id=task_id)
+    if request.method == 'POST':
+        task.completed = not task.completed  # Toggle status
+        task.save()
+        return redirect('index')
 
 def delete_task(request, task_id):
-    task = Task.objects.get(id=task_id)
-    task.delete()
-    return redirect('index')
+    task = get_object_or_404(Task, id=task_id)
+    if request.method == 'POST':
+        task.delete()
+        return redirect('index')
